requirements.txt
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
certifi==2022.6.15
decorator==5.1.1
Django==4.1
django-ipware==4.0.2
django-widget-tweaks==1.4.12
executing==0.10.0
h11==0.13.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
netifaces==0.11.0
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.9
pyshark==0.5.3
PySocks==1.7.1
scapy==2.5.0rc1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.4.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
wsproto==1.2.0
